ml_service/requirements.txt
fastapi
uvicorn
psycopg2-binary
pandas
torch
scikit-learn
python-dotenv
cachetools
